Script/requirements.txt
binance-futures-connector
pandas
matplotlib
scikit-learn
#python-binance